BaytesMCMC

Documentation for BaytesMCMC.

BaytesMCMC.AcceptStatistic

BaytesMCMC.Adaptionˢˡᵒʷ

BaytesMCMC.Adaptionᶠᵃˢᵗ

BaytesMCMC.DualAverage

BaytesMCMC.DualAverageParameter

BaytesMCMC.Exploration

BaytesMCMC.InitialStepsizeSearch

BaytesMCMC.MCMC

BaytesMCMC.MCMCConstructor

BaytesMCMC.MCMCDefault

BaytesMCMC.MCMCDiagnostics

BaytesMCMC.MCMCTune

BaytesMCMC.MatrixMetric

BaytesMCMC.MatrixTune

BaytesMCMC.PhaseTune

BaytesMCMC.Proposal

BaytesMCMC.SamplingPhase

BaytesMCMC.StepSizeTune

BaytesMCMC.Warmup

BaytesCore.adapt!

BaytesCore.infer

BaytesCore.infer

BaytesCore.init

BaytesCore.propose

BaytesCore.propose!

BaytesCore.results

BaytesCore.update

BaytesCore.update!

BaytesCore.update!

BaytesCore.update!

BaytesCore.update!

BaytesCore.update!

BaytesCore.update!

BaytesMCMC._infer_generated

BaytesMCMC.bisect_stepsize

BaytesMCMC.chain!

BaytesMCMC.find_crossing_stepsize

BaytesMCMC.find_initial_stepsize

BaytesMCMC.get_Σ

BaytesMCMC.print_divergences

BaytesMCMC.printtune

BaytesMCMC.AcceptStatistic — Type

struct AcceptStatistic{T<:AbstractFloat}

Stores in diagnostics.

Fields

rate::AbstractFloat

Acceptance rate

accepted::Bool

Step accepted or rejected

BaytesMCMC.Adaptionˢˡᵒʷ — Type

Explore typical set and tune MCMC parameter(initial bias rapidly vanishes and the estimators become much more accurate.)

BaytesMCMC.Adaptionᶠᵃˢᵗ — Type

Tune final stepsize for fixed Covariance Matrix.

BaytesMCMC.DualAverage — Type

struct DualAverage{T<:AbstractFloat}

Contains DualAverage tuning information and runtime parameter.

Fields

adaption::DualAverageParameter

μ::AbstractFloat

Upwards bias for target acceptance rate - proposals are biased upwards to stay away from 0.

t::Int64

Time update, starts with 0

H̄::AbstractFloat

Average part of first equation in Hoffman(2014), p 1607, (6)

logϵ::AbstractFloat

Log step

logϵ̄::AbstractFloat

AVERAGED log step

BaytesMCMC.DualAverageParameter — Type

struct DualAverageParameter{T<:AbstractFloat}

Contains information for default Dual Averaging algorithm.

Fields

δ::AbstractFloat

Target acceptance rate

γ::AbstractFloat

Regularization scale

κ::AbstractFloat

Relaxation exponent - for Average log step size

t₀::Int64

Offset

BaytesMCMC.Exploration — Type

Sample parameter with tuned MCMC (gradually reducing the precision error of the MCMC estimators towards zero.)

BaytesMCMC.InitialStepsizeSearch — Type

struct InitialStepsizeSearch{T<:AbstractFloat}

Initial Step Size Search factors.

Fields

a_min::AbstractFloat

Lowest local acceptance rate.

a_max::AbstractFloat

Highest local acceptance rate.

ϵ₀::AbstractFloat

Initial stepsize.

C::AbstractFloat

Scale factor for initial bracketing, > 1.

maxiter_crossing::Int64

Maximum number of iterations for initial bracketing.

maxiter_bisect::Int64

Maximum number of iterations for bisection.

BaytesMCMC.MCMC — Type

struct MCMC{M<:MCMCKernel, N<:MCMCTune} <: BaytesCore.AbstractAlgorithm

Stores information for proposal step.

Fields

kernel::MCMCKernel

MCMC sampler

tune::MCMCTune

Tuning configuration for kernel.

BaytesMCMC.MCMCConstructor — Type

Callable struct to make initializing MCMC sampler easier in sampling library.

Examples

BaytesMCMC.MCMCDefault — Type

struct MCMCDefault{K<:NamedTuple, S}

Default arguments for MCMC constructor.

Fields

config_kw::NamedTuple

Individual keyword arguments for tuning different MCMC engines.

GradientBackend::Any

Gradient backend used in MCMC step. Not used if Metropolis sampler is chosen.

TunedModel::Bool

Boolean if initial parameter are fixed or resampled.

generated::Bool

Boolean if generate(_rng, objective) for corresponding model is stored in MCMC Diagnostics.

BaytesMCMC.MCMCDiagnostics — Type

struct MCMCDiagnostics{R<:AbstractFloat, E<:MCMCKernelDiagnostics, T, G} <: BaytesCore.AbstractDiagnostics

MCMC Diagnostics container.

Fields

ℓθᵤ::AbstractFloat

Evaluation of target function at current iteration.

temperature::AbstractFloat

Temperature for log posterior evaluation

divergence::Bool

Boolean if diverged.

accept::AcceptStatistic

Acceptance Rate of current step.

sampler::MCMCKernelDiagnostics

Sampler specific diagnostics.

prediction::Any

Predicted sample of model.

generated::Any

Generated quantities specified for objective

iter::Int64

Current iteration number.

BaytesMCMC.MCMCTune — Type

struct MCMCTune{A<:UpdateBool, B<:UpdateBool, F<:AbstractFloat, T<:ModelWrappers.Tagged, E<:Tuple, P<:Proposal} <: BaytesCore.AbstractTune

MCMC Tuning container.

Fields

tagged::ModelWrappers.Tagged

Tagged Parameter.

phase::PhaseTune

Current Phase in MCMC Cycle

stepsize::StepSizeTune

Stepsize container

proposal::Proposal

Information for posterior covariance estimate

tempering::BaytesCore.TemperingTune

Information about tempering target function.

generated::Bool

Boolean if generated quantities should be generated while sampling

iter::BaytesCore.Iterator

Current iteration number

BaytesMCMC.MatrixMetric — Type

abstract type MatrixMetric

Choice for Posterior Covariance Matrix adaption.

Fields

BaytesMCMC.MatrixTune — Type

struct MatrixTune{A<:MatrixMetric}

Mass and Covariance Matrix specification for MCMC sampler, relevant for Euclidean Metric.

Fields

metric::MatrixMetric

Dense, Diagonal or Unit Mass Matrix.

shrinkage::Float64

Shrinkage parameter for Covariance estimation.

BaytesMCMC.PhaseTune — Type

struct PhaseTune{T<:Tuple}

Information about current SamplingPhase.

Fields

update::BaytesCore.Updater

Boolean if current iteration needs update.

iter::BaytesCore.Iterator

Counts current iteration in phase.

counter::BaytesCore.Iterator

MCMC Phases ~ counter = current cyle in slices/name/iterations.

slices::Vector{Int64}

Vector of MCMC iterations at each phase.

name::Tuple

Name of Sampling phases.

iterations::Vector{Int64}

Counts total iterations.

window::Vector{Int64}

Counts cycles in adaption phases Warmup(), Adaptionˢˡᵒʷ(), Adaptionᶠᵃˢᵗ(), i.e., 1-5-1 means 1 window init, 5 adapt, 1 exploration.

buffer::Vector{Int64}

Counts iteration in each cycle for adaption phases Warmup(), Adaptionˢˡᵒʷ(), Adaptionᶠᵃˢᵗ(), i.e., 50-25-50 means 1 time 50, 5 times 25*i, i in 1:5, 1 time 50.

BaytesMCMC.Proposal — Type

mutable struct Proposal{A<:UpdateBool, T<:AbstractFloat, P<:(AbstractMatrix), C<:(AbstractMatrix), M<:MatrixTune}

Proposal distribution container.

Fields

adaption::UpdateBool

Check if adaption true in current iteration

chain::Matrix{T} where T<:AbstractFloat

θᵤ samples in current MCMC Phase, used for Σ estimation

Σ::AbstractMatrix

Posterior Covariance estimate

Σ⁻¹ᶜʰᵒˡ::AbstractMatrix

Cholesky decomposition of Inverse Posterior Covariance matrix Σ, s.t. Σ⁻¹ᶜʰᵒˡ*Σ⁻¹ᶜʰᵒˡ'=Σ⁻¹. Used to generate random draws in HMC/NUTS sampler.

matrixtune::MatrixTune

Tuning parameter for Σ estimation

BaytesMCMC.SamplingPhase — Type

abstract type SamplingPhase

Sampling Container that determines tuning updates.

Fields

BaytesMCMC.StepSizeTune — Type

mutable struct StepSizeTune{A<:UpdateBool, F<:AbstractFloat}

Tuning stepsize parameter for MCMC algorithm.

Fields

adaption::UpdateBool

If true, stepsize will be adapted.

dualaverage::DualAverage

Dualaverage struct

ϵ::AbstractFloat

Current stepsize

BaytesMCMC.Warmup — Type

Find/converges to the typical set (estimators suffer from initial but ultimately transient biases)

BaytesCore.adapt! — Method

adapt!(proposal, metric)
adapt!(proposal)

Calculate regularized Covariance Matrix and Cholesky decomposition.

Examples

BaytesCore.infer — Method

infer(_rng, mcmc, model, data)

Infer type of predictions of MCMC sampler.

Examples

BaytesCore.infer — Method

infer(_rng, diagnostics, mcmc, model, data)

Infer MCMC diagnostics type.

Examples

BaytesCore.init — Method

init(, δ, ϵ; multiplier)

Initialize new DualAverage struct.

Examples

BaytesCore.propose! — Method

propose!(_rng, mcmc, model, data)
propose!(_rng, mcmc, model, data, update)

Inplace version of propose.

Examples

BaytesCore.propose — Method

propose(_rng, mcmc, objective)

Propose new parameter with mcmc sampler. If update=true, objective function will be updated with input model and data.

Examples

BaytesCore.results — Method

results(diagnosticsᵛ, mcmc, Ndigits, quantiles)

Print result for a single trace.

Examples

BaytesCore.update! — Method

update!(phase, iter)

Update current MCMC Phase.

Examples

BaytesCore.update! — Method

update!(tempering, adaption, phasename, iter)

Update Tempering with new temperature.

Examples

BaytesCore.update! — Method

update!(proposal, proposalupdate, θᵤ, phasename, phase)

Update Proposal according to current tuning phase.

Examples

BaytesCore.update! — Method

update!(proposal, phaseupdate, phasename, phase, θᵤ)

Update Proposal with new parameter θᵤ.

Examples

BaytesCore.update! — Method

update!(stepsize, stepsizeadaption, α, samplingphase, iterationupdate)

Update stepsize based on acceptance ratio α.

Examples

BaytesCore.update! — Method

update!(tune, result, acceptrate)

Update MCMC tuning fields at current iteration.

Examples

BaytesCore.update — Method

update(dualaverage, acceptratio)

Update DualAverage struct.

Examples

BaytesMCMC._infer_generated — Method

Infer type of generated quantities of MCMC sampler.

Examples

BaytesMCMC.bisect_stepsize — Function

bisect_stepsize(parameters, A, ϵ₀, ϵ₁)
bisect_stepsize(parameters, A, ϵ₀, ϵ₁, Aϵ₀)
bisect_stepsize(parameters, A, ϵ₀, ϵ₁, Aϵ₀, Aϵ₁)

Return the desired stepsize ϵ by bisection.

Examples

BaytesMCMC.chain! — Method

chain!(proposal, phasename, Niterations)

Assign new chain buffer with dedicated size.

Examples

BaytesMCMC.find_crossing_stepsize — Function

find_crossing_stepsize(parameters, A, ϵ₀)
find_crossing_stepsize(parameters, A, ϵ₀, Aϵ₀)

Find the stepsize for which the local acceptance rate A(ϵ) crosses a.

Examples

BaytesMCMC.find_initial_stepsize — Method

find_initial_stepsize(parameters, A)

Find an initial stepsize that matches the conditions of parameters (see InitialStepsizeSearch). A is the local acceptance ratio (unbounded).

Examples

BaytesMCMC.get_Σ — Method

get_Σ(Σ, shrinkage)

Calculate regularized Covariance Matrix.

Examples

BaytesMCMC.print_divergences — Method

print_divergences(diagnosticsᵛ, phase)

Print all divergences from a vector of MCMC Diagnostics.

Examples

BaytesMCMC.printtune — Function

printtune(tune)
printtune(tune, diagparam)

Some basic tuning output figures. Useful for debugging.

Examples