BaytesMCMC

Documentation for BaytesMCMC.

BaytesMCMC.DEFAULT_MAX_TREE_DEPTH

BaytesMCMC.MAX_DIRECTIONS_DEPTH

BaytesMCMC.REACHED_MAX_DEPTH

BaytesMCMC.Adaptionˢˡᵒʷ

BaytesMCMC.Adaptionᶠᵃˢᵗ

BaytesMCMC.ConfigCustom

BaytesMCMC.ConfigHMC

BaytesMCMC.ConfigMALA

BaytesMCMC.ConfigMetropolis

BaytesMCMC.ConfigNUTS

BaytesMCMC.ConfigProposal

BaytesMCMC.ConfigStepnumber

BaytesMCMC.ConfigStepsize

BaytesMCMC.ConfigTuningWindow

BaytesMCMC.Custom

BaytesMCMC.DiagnosticsCustom

BaytesMCMC.DiagnosticsHMC

BaytesMCMC.DiagnosticsMALA

BaytesMCMC.DiagnosticsMetropolis

BaytesMCMC.DiagnosticsNUTS

BaytesMCMC.Directions

BaytesMCMC.DualAverage

BaytesMCMC.DualAverageParameter

BaytesMCMC.EuclideanKineticEnergy

BaytesMCMC.Exploration

BaytesMCMC.GaussianKineticEnergy

BaytesMCMC.GaussianKineticEnergy

BaytesMCMC.GeneralizedTurnStatistic

BaytesMCMC.HMC

BaytesMCMC.Hamiltonian

BaytesMCMC.InitialStepsizeSearch

BaytesMCMC.InvalidTree

BaytesMCMC.KineticEnergy

BaytesMCMC.MALA

BaytesMCMC.MCMC

BaytesMCMC.MCMCConstructor

BaytesMCMC.MCMCDefault

BaytesMCMC.MCMCDiagnostics

BaytesMCMC.MCMCTune

BaytesMCMC.MatrixMetric

BaytesMCMC.MatrixTune

BaytesMCMC.Metropolis

BaytesMCMC.NUTS

BaytesMCMC.PhasePoint

BaytesMCMC.PhaseTune

BaytesMCMC.Proposal

BaytesMCMC.SamplingPhase

BaytesMCMC.StepNumberTune

BaytesMCMC.StepSizeTune

BaytesMCMC.TrajectoryCustom

BaytesMCMC.TrajectoryHMC

BaytesMCMC.TrajectoryMALA

BaytesMCMC.TrajectoryMetropolis

BaytesMCMC.TrajectoryNUTS

BaytesMCMC.Warmup

BaytesCore.generate_showvalues

BaytesCore.generate_showvalues

BaytesCore.generate_showvalues

BaytesCore.generate_showvalues

BaytesCore.generate_showvalues

BaytesCore.infer

BaytesCore.infer

BaytesCore.init

BaytesCore.init

BaytesCore.init

BaytesCore.init

BaytesCore.init

BaytesCore.init

BaytesCore.propagate

BaytesCore.propagate

BaytesCore.propagate

BaytesCore.propagate

BaytesCore.propagate

BaytesCore.propose

BaytesCore.propose!

BaytesCore.results

BaytesCore.results

BaytesCore.update

BaytesCore.update!

BaytesCore.update!

BaytesCore.update!

BaytesCore.update!

BaytesCore.update!

BaytesCore.update!

BaytesMCMC._is_turning

BaytesMCMC.acceptance_rate

BaytesMCMC.adjacent_tree

BaytesMCMC.biased_progressive_logprob2

BaytesMCMC.bisect_stepsize

BaytesMCMC.calculate_logprob2

BaytesMCMC.calculate_ρ♯

BaytesMCMC.chain!

BaytesMCMC.combine_proposals

BaytesMCMC.combine_turn_statistics

BaytesMCMC.combine_turn_statistics_in_direction

BaytesMCMC.combine_visited_statistics

BaytesMCMC.customdefaultkernel

BaytesMCMC.find_crossing_stepsize

BaytesMCMC.find_initial_stepsize

BaytesMCMC.get_acceptrate

BaytesMCMC.get_acceptrate

BaytesMCMC.get_Σ

BaytesMCMC.infer_generated

BaytesMCMC.is_turning

BaytesMCMC.leaf

BaytesMCMC.leaf_acceptance_statistic

BaytesMCMC.leapfrog

BaytesMCMC.move

BaytesMCMC.next_direction

BaytesMCMC.print_divergences

BaytesMCMC.printtune

BaytesMCMC.rand_bool_logprob

BaytesMCMC.rand_ρ

BaytesMCMC.sample_trajectory

BaytesMCMC.sample_tree

BaytesMCMC.ℓdensity

BaytesMCMC.∇K

BaytesMCMC.DEFAULT_MAX_TREE_DEPTH — Constant

Default maximum depth for trees.

BaytesMCMC.MAX_DIRECTIONS_DEPTH — Constant

Maximum number of iterations

BaytesMCMC.REACHED_MAX_DEPTH — Constant

Sentinel value for reaching maximum depth.

BaytesMCMC.Adaptionˢˡᵒʷ — Type

Explore typical set and tune MCMC parameter(initial bias rapidly vanishes and the estimators become much more accurate.)

BaytesMCMC.Adaptionᶠᵃˢᵗ — Type

Tune final stepsize for fixed Covariance Matrix.

BaytesMCMC.ConfigCustom — Type

Default Configuration for Custom sampler

BaytesMCMC.ConfigHMC — Type

struct ConfigHMC{K<:KineticEnergy, S<:ConfigStepnumber} <: BaytesCore.AbstractConfiguration

Default Configuration for HMC sampler.

Fields

δ::Float64

Target Acceptance Rate

window::ConfigTuningWindow

Default size for tuning iterations in each cycle.

energy::KineticEnergy

Kinetic Energy used in Hamiltonian: GaussianKineticEnergy

stepnumber::ConfigStepnumber

Step Number adaption

BaytesMCMC.ConfigMALA — Type

struct ConfigMALA <: BaytesCore.AbstractConfiguration

Default Configuration for MALA sampler.

Fields

δ::Float64

Target Acceptance Rate

window::ConfigTuningWindow

Default size for tuning iterations in each cycle.

BaytesMCMC.ConfigMetropolis — Type

Default Configuration for Metropolis sampler

BaytesMCMC.ConfigNUTS — Type

struct ConfigNUTS{K<:KineticEnergy} <: BaytesCore.AbstractConfiguration

Default Configuration for NUTS sampler.

Fields

δ::Float64

Target Acceptance Rate

window::ConfigTuningWindow

Default size for tuning iterations in each cycle.

energy::KineticEnergy

Kinetic Energy used in Hamiltonian: GaussianKineticEnergy

BaytesMCMC.ConfigProposal — Type

struct ConfigProposal{A<:UpdateBool, M<:MatrixMetric}

Default configuration for posterior Covariance Matrix adaption.

Fields

proposaladaption::UpdateBool

Boolean if Posterior covariance for proposal steps is adapted.

metric::MatrixMetric

Covariance estimate metric: MDense(), MDiagonal(), MUnit()

shrinkage::Float64

Shrinkage parameter towards Diagonal Matrix with equal variance

BaytesMCMC.ConfigStepnumber — Type

struct ConfigStepnumber{A<:UpdateBool}

Default configuration for stepsize adaption.

Fields

stepnumberadaption::UpdateBool

Step Number adaption

steps::Int64

Initial number of Steps

maxsteps::Int64

Maximal number of steps

∫dt::Float64

Desired Integration time

BaytesMCMC.ConfigStepsize — Type

struct ConfigStepsize{A<:UpdateBool, B<:UpdateBool}

Default configuration for stepsize adaption.

Fields

ϵ::Float64

Initial Discretization size

stepsizeadaption::UpdateBool

Step size adaption

initialstepsize::UpdateBool

Boolean if initial stepsize should be estimated

BaytesMCMC.ConfigTuningWindow — Type

struct ConfigTuningWindow

Default configuration for tuning iterations in each cycle.

Fields

window::Vector{Int64}

MCMC Phase tuning window lengths, i.e: 5 = 5 repeats of second window in phasenames

buffer::Vector{Int64}

(Increasing) Length of windows, i.e.: if window=3, buffer=10 -> total window length: 10-20-30

phasenames::Vector{SamplingPhase}

Name of phasenames, currently supported: Warmup(), Adaptionˢˡᵒʷ(), Adaptionᶠᵃˢᵗ(), Exploration()

BaytesMCMC.Custom — Type

mutable struct Custom{R<:ModelWrappers.ℓDensityResult, P} <: MCMCKernel

Custom algorithm container.

Fields

result::ModelWrappers.ℓDensityResult

Cached Result of last propagation step.

proposal::Any

Callable struct/closure/function of result.θᵤ that returns proposal distribution.

BaytesMCMC.DiagnosticsCustom — Type

struct DiagnosticsCustom <: MCMCKernelDiagnostics

Default Configuration for Custom sampler.

Fields

BaytesMCMC.DiagnosticsHMC — Type

struct DiagnosticsHMC{T<:AbstractFloat} <: MCMCKernelDiagnostics

Diagnostics fields for HMC sampler.

Fields

ϵ::AbstractFloat

Discretization size

steps::Int64

Number of Leapfrog steps

BaytesMCMC.DiagnosticsMALA — Type

struct DiagnosticsMALA{T<:AbstractFloat} <: MCMCKernelDiagnostics

Diagnostics for MALA sampler.

Fields

ϵ::AbstractFloat

Discretization size

BaytesMCMC.DiagnosticsMetropolis — Type

struct DiagnosticsMetropolis{T<:AbstractFloat} <: MCMCKernelDiagnostics

Default Configuration for Metropolis sampler.

Fields

ϵ::AbstractFloat

Discretization size

BaytesMCMC.DiagnosticsNUTS — Type

struct DiagnosticsNUTS{T<:AbstractFloat} <: MCMCKernelDiagnostics

Diagnostics fields for NUTS sampler.

Fields

ℓH::AbstractFloat

Log density (negative energy).

depth::Int64

Depth of the tree.

termination::InvalidTree

Reason for termination. See InvalidTree and REACHED_MAX_DEPTH.

acceptance_rate::AbstractFloat

Acceptance rate statistic.

ϵ::AbstractFloat

Discretization size

steps::Int64

Number of leapfrog steps evaluated.

directions::Directions

Directions for tree doubling (useful for debugging).

BaytesMCMC.Directions — Type

struct Directions

Internal type implementing random directions.

Fields

flags::UInt32

BaytesMCMC.DualAverage — Type

struct DualAverage{T<:AbstractFloat}

Contains DualAverage tuning information and runtime parameter.

Fields

adaption::DualAverageParameter

μ::AbstractFloat

Upwards bias for target acceptance rate - proposals are biased upwards to stay away from 0.

t::Int64

Time update, starts with 0

H̄::AbstractFloat

Average part of first equation in Hoffman(2014), p 1607, (6)

logϵ::AbstractFloat

Log step

logϵ̄::AbstractFloat

AVERAGED log step

BaytesMCMC.DualAverageParameter — Type

struct DualAverageParameter{T<:AbstractFloat}

Contains information for default Dual Averaging algorithm.

Fields

δ::AbstractFloat

Target acceptance rate

γ::AbstractFloat

Regularization scale

κ::AbstractFloat

Relaxation exponent - for Average log step size

t₀::Int64

Offset

BaytesMCMC.EuclideanKineticEnergy — Type

Gaussian Kinetic Energy, independent of position parameter.

BaytesMCMC.Exploration — Type

Sample parameter with tuned MCMC (gradually reducing the precision error of the MCMC estimators towards zero.)

BaytesMCMC.GaussianKineticEnergy — Type

mutable struct GaussianKineticEnergy{T<:(AbstractMatrix), S<:(AbstractMatrix)} <: EuclideanKineticEnergy

Gaussian kinetic energy, which is independent of q.

Fields

Σ::AbstractMatrix

Inverse Mass Matrix Σ ~ Posterior Covariance Matrix

Mᶜʰᵒˡ::AbstractMatrix

Cholesky decomposition of Mass matrix M, s.t. Mᶜʰᵒˡ*Mᶜʰᵒˡ'=M. Used to generate random draws

BaytesMCMC.GaussianKineticEnergy — Method

Gaussian kinetic energy with the given inverse covariance matrix Σ.

BaytesMCMC.GeneralizedTurnStatistic — Type

struct GeneralizedTurnStatistic{T}

Statistics for the identification of turning points. See Betancourt (2017, appendix), and subsequent discussion of improvements at https://discourse.mc-stan.org/t/nuts-misses-u-turns-runs-in-circles-until-max-treedepth/9727/. Momenta p₋ and p₊ are kept so that they can be added to ρ when combining turn statistics. Turn detection is always done by combine_turn_statistics, which returns nothing in case of turning. A GeneralizedTurnStatistic should always correspond to a trajectory that is not turning (or a leaf node, where the concept does not apply).

Fields

ρ₋::Any

momentum at the left edge of the trajectory

ρ♯₋::Any

p♯ at the left edge of the trajectory

ρ₊::Any

momentum at the right edge of the trajectory

ρ♯₊::Any

p♯ at the right edge of the trajectory

ρₛ::Any

sum of momenta along trajectory

BaytesMCMC.HMC — Type

mutable struct HMC{R<:ModelWrappers.ℓObjectiveResult, D<:ModelWrappers.AbstractDifferentiableTune, C<:KineticEnergy, T<:UpdateBool} <: MCMCKernel

HMC - Container used throughout sampling process.

Fields

result::ModelWrappers.ℓObjectiveResult

Target result stored for caching.

diff::ModelWrappers.AbstractDifferentiableTune

Differentiation tuning struct.

energy::KineticEnergy

Energy used for Hamiltonian.

stepnumber::StepNumberTune

Tuning struct for discretization steps

BaytesMCMC.Hamiltonian — Type

struct Hamiltonian{E<:KineticEnergy, D<:ModelWrappers.DiffObjective}

Hamiltonian struct that holds Kinetic energy and log target density function.

Fields

K::KineticEnergy

The kinetic energy specification.

objective::ModelWrappers.DiffObjective

Differentiable log target density of model, see ModelWrappers.

BaytesMCMC.InitialStepsizeSearch — Type

struct InitialStepsizeSearch{T<:AbstractFloat}

Initial Step Size Search factors.

Fields

a_min::AbstractFloat

Lowest local acceptance rate.

a_max::AbstractFloat

Highest local acceptance rate.

ϵ₀::AbstractFloat

Initial stepsize.

C::AbstractFloat

Scale factor for initial bracketing, > 1.

maxiter_crossing::Int64

Maximum number of iterations for initial bracketing.

maxiter_bisect::Int64

Maximum number of iterations for bisection.

BaytesMCMC.InvalidTree — Type

struct InvalidTree

Information about an invalid (sub)tree, using positions relative to the starting node.

When left < right, this tree was turning.

When left == right, this is a divergent node.

left == 1 && right == 0 is used as a sentinel value for reaching maximum depth without

encountering any invalid trees (see REACHED_MAX_DEPTH. All other left > right values are disallowed.

Fields

left::Int64

right::Int64

BaytesMCMC.KineticEnergy — Type

Kinetic Energy in HMC setting.

BaytesMCMC.MALA — Type

mutable struct MALA{M<:ModelWrappers.ℓGradientResult, D<:ModelWrappers.AbstractDifferentiableTune} <: MCMCKernel

MALA algorithm container.

Fields

result::ModelWrappers.ℓGradientResult

Cached Result of last propagation step.

diff::ModelWrappers.AbstractDifferentiableTune

Differentiation tuning container

BaytesMCMC.MCMC — Type

struct MCMC{M<:MCMCKernel, N<:MCMCTune} <: BaytesCore.AbstractAlgorithm

Stores information for proposal step.

Fields

kernel::MCMCKernel

MCMC sampler

tune::MCMCTune

Tuning configuration for kernel.

BaytesMCMC.MCMCConstructor — Type

Callable struct to make initializing MCMC sampler easier in sampling library.

Examples

BaytesMCMC.MCMCDefault — Type

struct MCMCDefault{K<:NamedTuple, S<:ConfigStepsize, P<:ConfigProposal, G, U<:UpdateBool}

Default arguments for MCMC constructor.

Fields

kernel::NamedTuple

Individual keyword arguments for tuning different MCMC kernels.

stepsize::ConfigStepsize

Stepsize default configuration for adaption.

proposal::ConfigProposal

Proposal distribution default configuration for adaption.

GradientBackend::Any

Gradient backend used in MCMC step. Not used if Metropolis sampler is chosen.

TunedModel::Bool

Boolean if initial parameter are fixed or resampled.

generated::UpdateBool

Boolean if generate(_rng, objective) for corresponding model is stored in MCMC Diagnostics.

BaytesMCMC.MCMCDiagnostics — Type

struct MCMCDiagnostics{P, K<:MCMCKernelDiagnostics, G} <: BaytesCore.AbstractDiagnostics

MCMC Diagnostics container.

Fields

base::BaytesCore.BaseDiagnostics

Diagnostics used for all Baytes kernels

kernel::MCMCKernelDiagnostics

Kernel specific diagnostics.

divergence::Bool

Boolean if diverged.

accept::BaytesCore.AcceptStatistic

Acceptance Rate of current step.

generated::Any

Generated quantities specified for objective

BaytesMCMC.MCMCTune — Type

struct MCMCTune{A<:UpdateBool, B<:UpdateBool, F<:AbstractFloat, T<:ModelWrappers.Tagged, E<:Tuple, P<:Proposal} <: BaytesCore.AbstractTune

MCMC Tuning container.

Fields

tagged::ModelWrappers.Tagged

Tagged Parameter.

phase::PhaseTune

Current Phase in MCMC Cycle

stepsize::StepSizeTune

Stepsize container

proposal::Proposal

Information for posterior covariance estimate

generated::UpdateBool

Boolean if generated quantities should be generated while sampling

iter::BaytesCore.Iterator

Current iteration number

BaytesMCMC.MatrixMetric — Type

abstract type MatrixMetric

Choice for Posterior Covariance Matrix adaption.

Fields

BaytesMCMC.MatrixTune — Type

struct MatrixTune{A<:MatrixMetric}

Mass and Covariance Matrix specification for MCMC sampler, relevant for Euclidean Metric.

Fields

metric::MatrixMetric

Dense, Diagonal or Unit Mass Matrix.

shrinkage::Float64

Shrinkage parameter for Covariance estimation.

BaytesMCMC.Metropolis — Type

mutable struct Metropolis{R<:ModelWrappers.ℓDensityResult} <: MCMCKernel

Metropolis algorithm container.

Fields

result::ModelWrappers.ℓDensityResult

Cached Result of last propagation step.

BaytesMCMC.NUTS — Type

mutable struct NUTS{R<:ModelWrappers.ℓObjectiveResult, D<:ModelWrappers.AbstractDifferentiableTune, C<:KineticEnergy} <: MCMCKernel

NUTS - Container used throughout sampling process.

Fields

result::ModelWrappers.ℓObjectiveResult

Target result stored for caching.

diff::ModelWrappers.AbstractDifferentiableTune

Differentiation tuning struct.

energy::KineticEnergy

Energy used for Hamiltonian.

BaytesMCMC.PhasePoint — Type

struct PhasePoint{T<:ModelWrappers.ℓObjectiveResult, S<:Real}

A point in phase space, consists of a position ModelWrappers.ℓObjectiveResult and a momentum ρ.

Fields

result::ModelWrappers.ℓObjectiveResult

ModelWrappers.ℓObjectiveResult container

ρ::Vector{S} where S<:Real

Momentum

BaytesMCMC.PhaseTune — Type

struct PhaseTune{T<:Tuple}

Information about current SamplingPhase.

Fields

update::BaytesCore.Updater

Boolean if current iteration needs update.

iter::BaytesCore.Iterator

Counts current iteration in phase.

counter::BaytesCore.Iterator

MCMC Phases ~ counter = current cyle in slices/name/iterations.

slices::Vector{Int64}

Vector of MCMC iterations at each phase.

name::Tuple

Name of Sampling phases.

iterations::Vector{Int64}

Counts total iterations.

window::Vector{Int64}

Counts cycles in adaption phases Warmup(), Adaptionˢˡᵒʷ(), Adaptionᶠᵃˢᵗ(), i.e., 1-5-1 means 1 window init, 5 adapt, 1 exploration.

buffer::Vector{Int64}

Counts iteration in each cycle for adaption phases Warmup(), Adaptionˢˡᵒʷ(), Adaptionᶠᵃˢᵗ(), i.e., 50-25-50 means 1 time 50, 5 times 25*i, i in 1:5, 1 time 50.

BaytesMCMC.Proposal — Type

mutable struct Proposal{A<:UpdateBool, T<:AbstractFloat, P<:(AbstractMatrix), C<:(AbstractMatrix), M<:MatrixTune}

Proposal distribution container.

Fields

adaption::UpdateBool

Check if adaption true in current iteration

chain::Matrix{T} where T<:AbstractFloat

θᵤ samples in current MCMC Phase, used for Σ estimation

Σ::AbstractMatrix

Posterior Covariance estimate

Σ⁻¹ᶜʰᵒˡ::AbstractMatrix

Cholesky decomposition of Inverse Posterior Covariance matrix Σ, s.t. Σ⁻¹ᶜʰᵒˡ*Σ⁻¹ᶜʰᵒˡ'=Σ⁻¹. Used to generate random draws in HMC/NUTS sampler.

matrixtune::MatrixTune

Tuning parameter for Σ estimation

BaytesMCMC.SamplingPhase — Type

abstract type SamplingPhase

Sampling Container that determines tuning updates.

Fields

BaytesMCMC.StepNumberTune — Type

mutable struct StepNumberTune{T<:UpdateBool}

Contains information for number of discretization steps in MCMC algorithm.

Fields

adaption::UpdateBool

If true, number of steps will be adapted.

steps::Int64

Number of discretization steps.

stepsᵐᵃˣ::Int64

Maximum number of discretization steps.

∫dt::Float64

Desired Integration time

BaytesMCMC.StepSizeTune — Type

mutable struct StepSizeTune{A<:UpdateBool, F<:AbstractFloat}

Tuning stepsize parameter for MCMC algorithm.

Fields

adaption::UpdateBool

If true, stepsize will be adapted.

dualaverage::DualAverage

Dualaverage struct

ϵ::AbstractFloat

Current stepsize

BaytesMCMC.TrajectoryCustom — Type

struct TrajectoryCustom{P, F<:ModelWrappers.ℓObjectiveResult}

Representation of a trajectory.

Fields

proposal::Any

Proposal distribution

result₀::ModelWrappers.ℓObjectiveResult

Log Target result at initial point, see ModelWrappersesTools.

BaytesMCMC.TrajectoryHMC — Type

struct TrajectoryHMC{S<:Hamiltonian, F<:AbstractFloat}

Representation of a trajectory, ie a Hamiltonian with a discrete integrator that also checks for divergence.

Fields

H::Hamiltonian

Hamiltonian.

ℓH₀::AbstractFloat

Log density of negative log energy at initial point.

ϵ::AbstractFloat

Stepsize for leapfrog.

BaytesMCMC.TrajectoryMALA — Type

struct TrajectoryMALA{S<:(AbstractMatrix), F<:ModelWrappers.ℓObjectiveResult, T<:AbstractFloat}

Representation of a trajectory.

Fields

Σ::AbstractMatrix

Proposal Covariance

result₀::ModelWrappers.ℓObjectiveResult

Log Target result at initial point, see ModelWrappers.

ϵ::AbstractFloat

Discretization size

BaytesMCMC.TrajectoryMetropolis — Type

struct TrajectoryMetropolis{S<:(AbstractMatrix), F<:ModelWrappers.ℓObjectiveResult, T<:AbstractFloat}

Representation of a trajectory.

Fields

Σ::AbstractMatrix

Proposal Covariance

result₀::ModelWrappers.ℓObjectiveResult

Log Target result at initial point, see ModelWrappersesTools.

ϵ::AbstractFloat

Discretization size

BaytesMCMC.TrajectoryNUTS — Type

Representation of a trajectory, ie a Hamiltonian with a discrete integrator that also checks for divergence.

BaytesMCMC.Warmup — Type

Find/converges to the typical set (estimators suffer from initial but ultimately transient biases)

BaytesCore.generate_showvalues — Method

generate_showvalues(diagnostics)

Show relevant diagnostic results.

Examples

BaytesCore.generate_showvalues — Method

generate_showvalues(diagnostics)

Show relevant diagnostic results.

Examples

BaytesCore.generate_showvalues — Method

generate_showvalues(diagnostics)

Show relevant diagnostic results.

Examples

BaytesCore.generate_showvalues — Method

generate_showvalues(diagnostics)

Show relevant diagnostic results.

Examples

BaytesCore.generate_showvalues — Method

generate_showvalues(diagnostics)

Show relevant diagnostic results.

Examples

BaytesCore.infer — Method

infer(_rng, mcmc, model, data)

Infer type of predictions of MCMC sampler.

Examples

BaytesCore.infer — Method

infer(_rng, diagnostics, mcmc, model, data)

Infer MCMC diagnostics type.

Examples

BaytesCore.init — Method

init(, mcmc, objective, proposalconfig; δ, window, proposal)

Initialize Custom custom configurations.

Examples

BaytesCore.init — Method

init(, mcmc, objective, proposalconfig; δ, window, energy, stepnumber)

Initialize HMC custom configurations.

Examples

BaytesCore.init — Method

init(, mcmc, objective, proposalconfig; δ, window)

Initialize Mala custom configurations.

Examples

BaytesCore.init — Method

init(, mcmc, objective, proposalconfig; δ, window)

Initialize Metropolis custom configurations.

Examples

BaytesCore.init — Method

init(, mcmc, objective, proposalconfig; δ, window, energy)

Initialize NUTS custom configurations.

Examples

BaytesCore.init — Method

init(, δ, ϵ; multiplier)

Initialize new DualAverage struct.

Examples

BaytesCore.propagate — Method

propagate Custom sampler forward

BaytesCore.propagate — Method

propagate(_rng, kernel, tune, objective)

Propagate forward one HMC step.

Examples

BaytesCore.propagate — Method

propagate(_rng, kernel, tune, objective)

Propagate forward one MALA step.

Examples

BaytesCore.propagate — Method

propagate Metropolis sampler forward

BaytesCore.propagate — Method

propagate(_rng, kernel, tune, objective)

Propagate forward one proposal step.

Examples

BaytesCore.propose! — Method

propose!(_rng, mcmc, model, data)
propose!(_rng, mcmc, model, data, temperature)
propose!(_rng, mcmc, model, data, temperature, update)

Inplace version of propose.

Examples

BaytesCore.propose — Method

propose(_rng, mcmc, objective)

Propose new parameter with mcmc sampler. If update=true, objective function will be updated with input model and data.

Examples

BaytesCore.results — Method

results(diagnosticsᵛ, mcmc, Ndigits, quantiles)

Print result for a single trace.

Examples

BaytesCore.results — Method

results(diagnostics, Ndigits, quantiles)

Summarize results of Vector of diagnostics.

Examples

BaytesCore.update! — Method

update!(phase, iter)

Update current MCMC Phase.

Examples

BaytesCore.update! — Method

update!(proposal, metric)
update!(proposal)

Calculate regularized Covariance Matrix and Cholesky decomposition.

Examples

BaytesCore.update! — Method

update!(proposal)
update!(proposal, proposalupdate, θᵤ, phasename, phase)

Update Proposal according to current tuning phase.

Examples

BaytesCore.update! — Method

update!(proposal)
update!(proposal, phaseupdate, phasename, phase, θᵤ)

Update Proposal with new parameter θᵤ.

Examples

BaytesCore.update! — Method

update!(stepsize, stepsizeadaption, α, samplingphase, iterationupdate)

Update stepsize based on acceptance ratio α.

Examples

BaytesCore.update! — Method

update!(tune, result, acceptrate)

Update MCMC tuning fields at current iteration.

Examples

BaytesCore.update — Method

update(dualaverage, acceptratio)

Update DualAverage struct.

Examples

BaytesMCMC._is_turning — Method

Internal test for turning. See Betancourt (2017, appendix).

BaytesMCMC.acceptance_rate — Method

acceptance_rate(A)

Return the acceptance rate (a Real betwen 0 and 1).

Examples

BaytesMCMC.adjacent_tree — Method

adjacent_tree(_rng, trajectory, z, i, depth, is_forward)

Traverse the tree of given depth adjacent to point z in trajectory. is_forward specifies the direction, rng is used for random numbers in combine_proposals. i is an integer position relative to the initial node (0). The first value is either

an InvalidTree, indicating the first divergent node or turning subtree that was

encounteted and invalidated this tree.

a tuple of `(ζ, ω, τ, z′, i′), with

ζ: the proposal from the tree.

ω: the log weight of the subtree that corresponds to the proposal

τ: turn statistics

z′: the last node of the tree

i′: the position of the last node relative to the initial node.

The second value is always the visited node statistic.

Examples

BaytesMCMC.biased_progressive_logprob2 — Method

biased_progressive_logprob2(bias, ω₁, ω₂)
biased_progressive_logprob2(bias, ω₁, ω₂, ω)

Given (relative) log probabilities ω₁ and ω₂, return the log probabiliy of drawing a sample from the second (logprob2). When bias, biases towards the second argument, introducing anti-correlations.

Examples

BaytesMCMC.bisect_stepsize — Function

bisect_stepsize(parameters, A, ϵ₀, ϵ₁)
bisect_stepsize(parameters, A, ϵ₀, ϵ₁, Aϵ₀)
bisect_stepsize(parameters, A, ϵ₀, ϵ₁, Aϵ₀, Aϵ₁)

Return the desired stepsize ϵ by bisection.

Examples

BaytesMCMC.calculate_logprob2 — Method

calculate_logprob2(_, is_doubling, ω₁, ω₂, ω)

Calculate the log probability if selecting the subtree corresponding to ω₂. Being the log of a probability, it is always ≤ 0, but implementations are allowed to return and accept values > 0 and treat them as 0. When is_doubling, the tree corresponding to ω₂ was obtained from a doubling step (this can be relevant eg for biased progressive sampling). The value ω = logaddexp(ω₁, ω₂) is provided for avoiding redundant calculations. See biased_progressive_logprob2 for an implementation.

Examples

BaytesMCMC.calculate_ρ♯ — Function

Return p♯ = M⁻¹⋅p, used for turn diagnostics.

BaytesMCMC.chain! — Method

chain!(proposal, phasename, Niterations)

Assign new chain buffer with dedicated size.

Examples

BaytesMCMC.combine_proposals — Method

combine_proposals(_rng, trajectory, z₁, z₂, logprob2, is_forward)

Combine two proposals ζ₁, ζ₂ on trajectory, with log probability logprob2 for selecting ζ₂. ζ₁ is before ζ₂ iff is_forward.

Examples

BaytesMCMC.combine_turn_statistics — Method

combine_turn_statistics(trajectory, x, y)

Combine turn statistics on trajectory. Implementation can assume that the trees that correspond to the turn statistics have the same ordering. When

τ = combine_turn_statistics(trajectory, τ₁, τ₂)
is_turning(trajectory, τ)

the combined turn statistic τ is guaranteed not to escape the caller, so it can eg change type.

Examples

BaytesMCMC.combine_turn_statistics_in_direction — Method

combine_turn_statistics_in_direction(trajectory, τ₁, τ₂, is_forward)

Combine turn statistics with the given direction. When is_forward, τ₁ is before τ₂, otherwise after. Internal helper function.

Examples

BaytesMCMC.combine_visited_statistics — Method

combine_visited_statistics(_, v, w)

Combine visited node statistics for adjacent trees trajectory. Implementation should be invariant to the ordering of v₁ and v₂ (ie the operation is commutative).

Examples

BaytesMCMC.customdefaultkernel — Method

customdefaultkernel(θ)

A placeholder function to configure a custom proposal distribution. Can also be a closure that captures, i.e., Σ., or a a callable struct:

Examples

#Mixture via callable struct
struct MyMixture{W, F, T}
    weights::W
    Σ1::F
    Σ2::T
end
function (mixture::MyMixture)(θ)
    @unpack weights, Σ1, Σ2 = mixture
    return MixtureModel([MvNormal(θ, Σ1), MvNormal(θ, Σ2)], weights)
end

#Independence sampler via callable struct without extra allocation
struct MyIndependenceSampler{W}
    dist::W
end
function (sampler::MyIndependenceSampler)(θ)
    return sampler.dist
end

BaytesMCMC.find_crossing_stepsize — Function

find_crossing_stepsize(parameters, A, ϵ₀)
find_crossing_stepsize(parameters, A, ϵ₀, Aϵ₀)

Find the stepsize for which the local acceptance rate A(ϵ) crosses a.

Examples

BaytesMCMC.find_initial_stepsize — Method

find_initial_stepsize(parameters, A)

Find an initial stepsize that matches the conditions of parameters (see InitialStepsizeSearch). A is the local acceptance ratio (unbounded).

Examples

BaytesMCMC.get_acceptrate — Method

get_acceptrate(_rng, kernel, objective, Σ)

Tune initial step size.

Examples

BaytesMCMC.get_acceptrate — Method

get_acceptrate(_rng, kernel, objective, Σ)

Tune initial step size.

Examples

BaytesMCMC.get_Σ — Method

get_Σ(Σ, shrinkage)

Calculate regularized Covariance Matrix.

Examples

BaytesMCMC.infer_generated — Method

Infer type of generated quantities of MCMC sampler.

Examples

BaytesMCMC.is_turning — Method

Internal test for turning. See Betancourt (2017, appendix).

BaytesMCMC.leaf — Method

leaf(trajectory, phasepoint, is_initial)

Information for a tree made of a single node. When is_initial == true, this is the first node. The first value is either

nothing for a divergent node,

a tuple containing the proposal ζ, the log weight (probability) of the node ω, the

turn statistics τ (never tested with is_turning for leaf nodes). The second value is the visited node information.

Examples

BaytesMCMC.leaf_acceptance_statistic — Method

leaf_acceptance_statistic(Δ, is_initial)

Acceptance statistic for a leaf. The initial leaf is considered not to be visited.

Examples

BaytesMCMC.leapfrog — Method

leapfrog(H, phasepoint, ϵ)

Take a leapfrog step of length ϵ from phasepoint along the Hamiltonian H.

Examples

BaytesMCMC.move — Method

move(trajectory, phasepoint, fwd)

Move along the trajectory in the specified direction. Return the new position.

Examples

BaytesMCMC.next_direction — Method

next_direction(directions)

Return the next direction flag and the new state of directions. Results are undefined for more than MAX_DIRECTIONS_DEPTH updates.

Examples

BaytesMCMC.print_divergences — Method

print_divergences(diagnosticsᵛ, phase)

Print all divergences from a vector of MCMC Diagnostics.

Examples

BaytesMCMC.printtune — Function

printtune(tune)
printtune(tune, diagparam)

Some basic tuning output figures. Useful for debugging.

Examples

BaytesMCMC.rand_bool_logprob — Method

rand_bool_logprob(_rng, logprob)

Random boolean which is true with the given probability exp(logprob), which can be ≥ 1 in which case no random value is drawn.

Examples

BaytesMCMC.rand_ρ — Function

Generate a random momentum from a kinetic energy at position ρ.

BaytesMCMC.sample_trajectory — Method

sample_trajectory(_rng, trajectory, z, directions)

Sample a trajectory starting at z, up to max_depth. directions determines the tree expansion directions. Return the following values

ζ: proposal from the tree

v: visited node statistics

termination: an InvalidTree (this includes the last doubling step turning, which is technically a valid tree) or REACHED_MAX_DEPTH when all subtrees were valid and no turning happens.

depth: the depth of the tree that was sampled from. Doubling steps that lead to an invalid adjacent tree do not contribute to depth.

Examples

BaytesMCMC.sample_tree — Method

sample_tree(_rng, trajectory, phasepoint; directions)

No-U-turn Hamiltonian Monte Carlo transition, using Hamiltonian H, starting at evaluated log density position Q, using stepsize ϵ. Parameters of kernel govern the details of tree construction. Return two values, the new evaluated log density position, and tree statistics.

Examples

BaytesMCMC.ℓdensity — Method

ℓdensity(H, phasepoint)

Log density for Hamiltonian H at phasepoint. If ℓ(q) == -Inf (rejected), skips the kinetic energy calculation.

Examples

BaytesMCMC.∇K — Function

Calculate the gradient of the logarithm of kinetic energy in momentum ρ.